Refactor virt-controller migration metrics
This PR refactors the migration metrics from KubeVirt virt-controller component by:

- placing them in the corresponding pkg/monitoring/metrics/virt-controller/ directory
- using a stricter style for metric declaration and organization
- uniformize metric usage with other components
