--- pkg/monitoring/migration/migration_suite_test.go
@@ -1,11 +1,0 @@
-package migration_test
-
-import (
-	"testing"
-
-	"kubevirt.io/client-go/testutils"
-)
-
-func TestMigration(t *testing.T) {
-	testutils.KubeVirtTestSuiteSetup(t)
-}
